backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
websockets==13.1
anthropic==0.42.0
python-dotenv==1.0.1
edge-tts
google-cloud-texttospeech
